feat: new component -- custom box, modify startpage with custom box
