+++ packages/frontend/src/app.scss
@@ -36,3 +36,4 @@
 @import './sass/components/textEditor';
 @import './sass/components/voteBox';
 @import './sass/components/writingField';
+@import './sass/components/customBox';
