+++ packages/frontend/src/sass/components/_customBox.scss
@@ -1,0 +1,149 @@
+.custom-box-container {
+    position: fixed;
+    top: 0;
+    z-index: 15;
+    width: 100%;
+    height: 100%;
+
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-position: center;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    .top-buttons {
+        position: fixed;
+        top: 0;
+        z-index: 16;
+        width: 100%;
+        padding: 16px 48px;
+
+        display: flex;
+        justify-content: space-between;
+
+        img {
+            width: 48px;
+            height: 48px;
+            cursor: pointer;
+        }
+    }
+
+    .box {
+        position: relative;
+        width: 609px;
+        height: 658px;
+        padding: 48px 32px;
+        overflow-y: scroll;
+
+        border-radius: 8px;
+        border: $grey-100 solid 6px;
+
+        text-align: center;
+
+        @media (max-width: 609px) {
+            width: 100%;
+        }
+    }
+
+    p {
+        font-size: $text-md;
+
+        strong {
+            cursor: pointer;
+        }
+    }
+
+    h1 {
+        font-size: $title-lg;
+        font-weight: 700;
+    }
+
+    h2 {
+        font-size: $title-md;
+        font-weight: 600;
+    }
+
+    h3 {
+        font-size: $title-sm;
+        font-weight: 600;
+    }
+
+    .box-dark {
+        background-color: $grey-800;
+        color: $grey-100;
+    }
+
+    .box-light {
+        background-color: $grey-100;
+        color: $black;
+
+        p {
+            text-align: center;
+        }
+    }
+
+    .box-buttons {
+        background-color: transparent;
+
+        button {
+            width: 100%;
+            font-size: $text-md;
+            font-weight: 600;
+            border-radius: 8px;
+            padding: 16px 24px;
+
+            &:hover {
+                background-color: $grey-300;
+            }
+
+            img {
+                width: 48px;
+                height: 48px;
+            }
+        }
+
+        .button-dark {
+            background-color: $black;
+            color: $white;
+            border: $white solid 1px;
+        }
+
+        .button-light {
+            background-color: $grey-100;
+            color: $black;
+            border: $black solid 1px;
+        }
+    }
+
+    .buttons-horizontal {
+        display: flex;
+        flex-direction: horizontal;
+        gap: 16px;
+    }
+
+    .buttons-bottom {
+        position: absolute;
+        bottom: 48px;
+        width: 100%;
+        padding-right: 64px;
+    }
+
+    .gap {
+        height: 24px;
+    }
+
+    textarea {
+        height: 172px;
+        border-radius: 8px;
+        border: $grey-200 1px solid;
+        padding: 8px 16px;
+    }
+
+    input {
+        width: 100%;
+        height: 60px;
+        padding: 8px 16px;
+    }
+}
